refactor: extract proposal list section for easier reuse
